add gun up level

+++ lib/system/system_config.dart
@@ -1,7 +1,7 @@
 class SystemConfig {
   static final double TIME_AMMO_PLAYER = 1.5;
   static final double LEVEL_GUN = 1;
-  static final double HEART_PLAYER = 3;
+  static final int HEART_PLAYER = 3;
   static final double HEART_BOSS = 1000;
   static final double DAMGE_GUN = 5;
   static final double TIME_ITEM_BOSS = 5;
